feat: 최신 뉴스의 상수 폴더 분리
